requirements.txt
tavily-python==0.7.14
qdrant-client==1.15.1
fastembed==0.7.4
python-dotenv==1.1.1
langchain==1.0.3
langchain-community==0.4.1
langchain-qdrant==1.1.0
langchain-text-splitters==1.0.0
langchain-core==1.1.1
arxiv==2.3.1
pypdf==6.4.0
langchain-huggingface==1.1.0
langchain-openai==1.1.0
onnxruntime-directml==1.23.0
fastapi==0.120.4
uvicorn==0.38.0
Jinja2==3.1.6
python-multipart==0.0.20
pydantic>=2.0.0
